Update interface.py
Updated Change_mode(args) function so that it would not only change which button is in mint, but also which set of coordinates is shown. Also changed the grid value of all J1/J2/J3 objects so that they would appear in the of their X/Y/Z counterparts when visible. Finally, added a variable which indicates whether we are in Joint mode or cartesian mode.